fixed timeout issue

+++ nemo_skills/evaluation/evaluator.py
@@ -51,7 +51,7 @@ class MathEvaluatorConfig:
     num_parallel_requests: int = 100
     in_memory_lines: int = 1500
     numeric_precision: int = 15
-    timeout: float = 10.0
+    timeout: int = 10
     ignore_cache: bool = False
     # if True will not attempt to re-extract based on \boxed or regex
     use_predicted_answer_key: bool = False
